// src/app/models/position.ts
export class Position{
    
       id: number;
       deviceId:number;
       protocol:string;
       deviceTime:any;
       fixTime:any;
       serverTime:any;
       outdated: boolean
       valid: boolean;
       latitude: any;
       longitude: any;
       altitude: any;
       speed: any;
       course: any;
       address:string;
       accuracy: any;
       network: { };
       attributes: { }
        
}